Cross-check possible ROMs for serial numbers
This will shorten the list of available games to just those that both
fit the extension requirements and have a serial number. While it may
not cover every game, there is an extremely high probability at least
one game in a folder has a serial. This assumes all games are obtained
legally.

Conflicts:
	Source/ui_android/java/com/virtualapplications/play/GameInfo.java