Reset known memory cards when IOP is reset.
Some games call Init after doing some MC operations (Zombie Attack) and get confused if known memory cards are reset there.
Gran Turismo 4 needs known memory card status to be reset at some point in its init process, but that game resets the IOP a few times, so, we can use that to reset that info.
Would probably be best to reset when the module is loaded rather than doing it when IOP is reset.